Added message identification mechanic
Added a listener_handler function that runs a thread which will collect
all incoming messages to a process and check their source process id
and message id. The receive methods will then only accept incoming
messages with the matching soruce process and message id.
This is necessary for the correct handling of multiple incoming
messages at the same time and with the use of async functions.